+++ src/app/api/rest/post.ts
@@ -1,3 +1,5 @@
+import { ApiResult, ApiResultType, ApiError, headers } from "."
+
 export interface Post {
     id?: number
     title: string
@@ -10,39 +12,9 @@ export interface PostOk {
     author: string
 }
 
-export interface PostNg {}
-
-export interface ApiError extends Error {}
-
-export enum ApiResultType {
-    OK = "ok",
-    NG= "ng",
-    ERROR= "error",
-}
-
-export type ApiResult<Ok, Ng, Error> =
-    | {
-        type: ApiResultType.OK
-        body: Ok
-    }
-    | {
-        type: ApiResultType.NG
-        body: Ng
-    }
-    | {
-        type: ApiResultType.ERROR
-        body: Error
-    }
-
-export type PostResult = ApiResult<PostOk, PostNg, ApiError>
+export type PostResult = ApiResult<PostOk, null, ApiError>
 
 export type GetsResult = ApiResult<Post[], null, ApiError>
-
-const headers = {
-    "X-Requested-With": "XMLHttpRequest",
-    "Accept": "application/json",
-    "Content-Type": "application/json",
-}
 
 export const post = (post: Post) => (
     fetch('http://localhost:3000/api/posts', {
@@ -60,3 +32,9 @@ export const gets = () => (
         headers,
     })
 )
+
+export {
+    ApiError,
+    ApiResult,
+    ApiResultType,
+}
